requirements.txt
langchain>=0.1.0
langchain_openai>=0.0.5
crewai>=0.28.0
chromadb>=0.4.22
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
tiktoken>=0.5.0
openai>=1.10.0